Convert Tabs component from .js to .tsx
